chore: migrate sbf tests
